iss-907: added ajc support simple fallback
