+++ hystrix-contrib/hystrix-javanica/src/main/java/com/netflix/hystrix/contrib/javanica/command/AbstractHystrixCommandFactory.java
@@ -22,6 +22,7 @@ import com.netflix.hystrix.HystrixCollapser;
 import com.netflix.hystrix.contrib.javanica.annotation.HystrixCommand;
 import com.netflix.hystrix.contrib.javanica.annotation.HystrixProperty;
 import com.netflix.hystrix.contrib.javanica.conf.HystrixPropertiesManager;
+import com.netflix.hystrix.contrib.javanica.utils.EnvUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.Validate;
 
@@ -32,6 +33,8 @@ import java.util.Map;
 
 import static com.netflix.hystrix.contrib.javanica.cache.CacheInvocationContextFactory.createCacheRemoveInvocationContext;
 import static com.netflix.hystrix.contrib.javanica.cache.CacheInvocationContextFactory.createCacheResultInvocationContext;
+import static com.netflix.hystrix.contrib.javanica.utils.EnvUtils.isCompileWeaving;
+import static com.netflix.hystrix.contrib.javanica.utils.ajc.AjcUtils.getOriginalMethod;
 
 /**
  * Base implementation of {@link HystrixCommandFactory} interface.
@@ -91,6 +94,8 @@ public abstract class AbstractHystrixCommandFactory<T extends AbstractHystrixCom
                     MetaHolder fmMetaHolder = MetaHolder.builder()
                             .obj(metaHolder.getObj())
                             .method(fallbackMethod)
+                            .ajcMethod(getAjcMethod(metaHolder.getObj(), fallbackMethod))
+                            .weavingMode(EnvUtils.getWeavingMode())
                             .args(metaHolder.getArgs())
                             .defaultCollapserKey(metaHolder.getDefaultCollapserKey())
                             .defaultCommandKey(fallbackMethod.getName())
@@ -99,7 +104,15 @@ public abstract class AbstractHystrixCommandFactory<T extends AbstractHystrixCom
                             .hystrixCommand(fallbackMethod.getAnnotation(HystrixCommand.class)).build();
                     fallbackAction = new LazyCommandExecutionAction(GenericHystrixCommandFactory.getInstance(), fmMetaHolder, collapsedRequests);
                 } else {
-                    fallbackAction = new MethodExecutionAction(metaHolder.getObj(), fallbackMethod, metaHolder.getArgs(), metaHolder);
+                    // if falback methid isn't annotated with command annotation then we don't need to get ajc method for this
+                    MetaHolder fmMetaHolder = MetaHolder.builder()
+                            .obj(metaHolder.getObj())
+                            .method(fallbackMethod)
+                            .ajcMethod(null)
+                            .weavingMode(EnvUtils.getWeavingMode())
+                            .args(metaHolder.getArgs()).build();
+
+                    fallbackAction = new MethodExecutionAction(fmMetaHolder.getObj(), fallbackMethod, fmMetaHolder.getArgs(), fmMetaHolder);
                 }
             } catch (NoSuchMethodException e) {
                 throw Throwables.propagate(e);
@@ -110,6 +123,13 @@ public abstract class AbstractHystrixCommandFactory<T extends AbstractHystrixCom
 
     abstract T create(HystrixCommandBuilder hystrixCommandBuilder);
 
+
+    private Method getAjcMethod(Object target, Method fallback) {
+        if (isCompileWeaving()) {
+            return getOriginalMethod(target.getClass(), fallback);
+        }
+        return null;
+    }
 
     private CommandAction createCacheKeyAction(MetaHolder metaHolder) {
         CommandAction cacheKeyAction = null;
