+++ hystrix-contrib/hystrix-javanica/src/test/java/com/netflix/hystrix/contrib/javanica/test/common/CommonUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.hystrix.contrib.javanica;
+package com.netflix.hystrix.contrib.javanica.test.common;
 
 
 import com.google.common.base.Function;
